+++ internal/common/basic.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	appsv1 "k8s.io/api/apps/v1"
+	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
@@ -54,7 +55,7 @@ func runBasic() {
 			wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 			Expect(err).NotTo(HaveOccurred())
 
-			Eventually(wait.Consistent(CheckControlPlaneNodesReady(wcClient, int(replicas)), 12, 5*time.Second)).
+			Eventually(wait.Consistent(CheckControlPlaneNodesReady(state.GetContext(), wcClient, int(replicas)), 12, 5*time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
@@ -68,35 +69,42 @@ func runBasic() {
 			wcClient, err := state.GetFramework().WC(state.GetCluster().Name)
 			Expect(err).NotTo(HaveOccurred())
 
-			Eventually(wait.Consistent(CheckWorkerNodesReady(wcClient, values), 12, 5*time.Second)).
+			Eventually(wait.Consistent(CheckWorkerNodesReady(state.GetContext(), wcClient, values), 12, 5*time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all its Deployments Ready (means all replicas are running)", func() {
-			Eventually(wait.Consistent(CheckAllDeploymentsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllDeploymentsReady(state.GetContext(), wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all its StatefulSets Ready (means all replicas are running)", func() {
-			Eventually(wait.Consistent(CheckAllStatefulSetsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllStatefulSetsReady(state.GetContext(), wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all its DaemonSets Ready (means all daemon pods are running)", func() {
-			Eventually(wait.Consistent(CheckAllDaemonSetsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllDaemonSetsReady(state.GetContext(), wcClient), 10, time.Second)).
+				WithTimeout(15 * time.Minute).
+				WithPolling(wait.DefaultInterval).
+				Should(Succeed())
+		})
+
+		It("has all its Jobs completed successfully", func() {
+			Eventually(wait.Consistent(CheckAllJobsSucceeded(state.GetContext(), wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all of its Pods in the Running state", func() {
-			Eventually(wait.Consistent(CheckAllPodsSuccessfulPhase(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllPodsSuccessfulPhase(state.GetContext(), wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
@@ -115,13 +123,13 @@ func runBasic() {
 			mcClient := state.GetFramework().MC()
 			cluster := state.GetCluster()
 
-			machinePools, err := state.GetFramework().GetMachinePools(context.Background(), cluster.Name, cluster.GetNamespace())
+			machinePools, err := state.GetFramework().GetMachinePools(state.GetContext(), cluster.Name, cluster.GetNamespace())
 			Expect(err).NotTo(HaveOccurred())
 			if len(machinePools) == 0 {
 				Skip("Machine pools are not found")
 			}
 
-			Eventually(wait.Consistent(CheckMachinePoolsReadyAndRunning(mcClient, cluster.Name, cluster.GetNamespace()), 5, 5*time.Second)).
+			Eventually(wait.Consistent(CheckMachinePoolsReadyAndRunning(state.GetContext(), mcClient, cluster.Name, cluster.GetNamespace()), 5, 5*time.Second)).
 				WithTimeout(30 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
@@ -129,8 +137,8 @@ func runBasic() {
 	})
 }
 
-func CheckControlPlaneNodesReady(wcClient *client.Client, expectedNodes int) func() error {
-	controlPlaneFunc := wait.AreNumNodesReady(context.Background(), wcClient, expectedNodes, &cr.MatchingLabels{"node-role.kubernetes.io/control-plane": ""})
+func CheckControlPlaneNodesReady(ctx context.Context, wcClient *client.Client, expectedNodes int) func() error {
+	controlPlaneFunc := wait.AreNumNodesReady(ctx, wcClient, expectedNodes, &cr.MatchingLabels{"node-role.kubernetes.io/control-plane": ""})
 
 	return func() error {
 		ok, err := controlPlaneFunc()
@@ -141,7 +149,7 @@ func CheckControlPlaneNodesReady(wcClient *client.Client, expectedNodes int) fun
 	}
 }
 
-func CheckWorkerNodesReady(wcClient *client.Client, values *application.ClusterValues) func() error {
+func CheckWorkerNodesReady(ctx context.Context, wcClient *client.Client, values *application.ClusterValues) func() error {
 	minNodes := 0
 	maxNodes := 0
 	for _, pool := range values.NodePools {
@@ -159,7 +167,7 @@ func CheckWorkerNodesReady(wcClient *client.Client, values *application.ClusterV
 		Max: maxNodes,
 	}
 
-	workersFunc := wait.AreNumNodesReadyWithinRange(context.Background(), wcClient, expectedNodes, client.DoesNotHaveLabels{"node-role.kubernetes.io/control-plane"})
+	workersFunc := wait.AreNumNodesReadyWithinRange(ctx, wcClient, expectedNodes, client.DoesNotHaveLabels{"node-role.kubernetes.io/control-plane"})
 
 	return func() error {
 		ok, err := workersFunc()
@@ -174,10 +182,10 @@ func CheckWorkerNodesReady(wcClient *client.Client, values *application.ClusterV
 	}
 }
 
-func CheckAllPodsSuccessfulPhase(wcClient *client.Client) func() error {
+func CheckAllPodsSuccessfulPhase(ctx context.Context, wcClient *client.Client) func() error {
 	return func() error {
 		podList := &corev1.PodList{}
-		err := wcClient.List(context.Background(), podList)
+		err := wcClient.List(ctx, podList)
 		if err != nil {
 			return err
 		}
@@ -195,10 +203,10 @@ func CheckAllPodsSuccessfulPhase(wcClient *client.Client) func() error {
 	}
 }
 
-func CheckAllDeploymentsReady(wcClient *client.Client) func() error {
+func CheckAllDeploymentsReady(ctx context.Context, wcClient *client.Client) func() error {
 	return func() error {
 		deploymentList := &appsv1.DeploymentList{}
-		err := wcClient.List(context.Background(), deploymentList)
+		err := wcClient.List(ctx, deploymentList)
 		if err != nil {
 			return err
 		}
@@ -217,10 +225,10 @@ func CheckAllDeploymentsReady(wcClient *client.Client) func() error {
 	}
 }
 
-func CheckAllStatefulSetsReady(wcClient *client.Client) func() error {
+func CheckAllStatefulSetsReady(ctx context.Context, wcClient *client.Client) func() error {
 	return func() error {
 		statefulSetList := &appsv1.StatefulSetList{}
-		err := wcClient.List(context.Background(), statefulSetList)
+		err := wcClient.List(ctx, statefulSetList)
 		if err != nil {
 			return err
 		}
@@ -239,10 +247,39 @@ func CheckAllStatefulSetsReady(wcClient *client.Client) func() error {
 	}
 }
 
-func CheckAllDaemonSetsReady(wcClient *client.Client) func() error {
+func CheckAllJobsSucceeded(ctx context.Context, wcClient *client.Client) func() error {
+	return func() error {
+		jobList := &batchv1.JobList{}
+		err := wcClient.List(ctx, jobList)
+		if err != nil {
+			return err
+		}
+
+		var loopErr error
+		for _, job := range jobList.Items {
+			if job.Status.Succeeded == 0 {
+				logger.Log("Job %s/%s has not succeeded. (Failed: '%d')", job.ObjectMeta.Namespace, job.ObjectMeta.Name, job.Status.Failed)
+				// We wrap the erros so that we can log out for all failures, not just the first found
+				if loopErr != nil {
+					loopErr = fmt.Errorf("%w, job %s/%s has not succeeded", loopErr, job.ObjectMeta.Namespace, job.ObjectMeta.Name)
+				} else {
+					loopErr = fmt.Errorf("job %s/%s has not succeeded", job.ObjectMeta.Namespace, job.ObjectMeta.Name)
+				}
+			}
+		}
+		if loopErr != nil {
+			return loopErr
+		}
+
+		logger.Log("All (%d) Jobs have completed successfully", len(jobList.Items))
+		return nil
+	}
+}
+
+func CheckAllDaemonSetsReady(ctx context.Context, wcClient *client.Client) func() error {
 	return func() error {
 		daemonSetList := &appsv1.DaemonSetList{}
-		err := wcClient.List(context.Background(), daemonSetList)
+		err := wcClient.List(ctx, daemonSetList)
 		if err != nil {
 			return err
 		}
@@ -263,7 +300,7 @@ func CheckAllDaemonSetsReady(wcClient *client.Client) func() error {
 
 // CheckMachinePoolsReadyAndRunning checks if all MachinePool resources have Ready condition with Status True and are in
 // Running phase.
-func CheckMachinePoolsReadyAndRunning(mcClient *client.Client, clusterName string, clusterNamespace string) func() error {
+func CheckMachinePoolsReadyAndRunning(ctx context.Context, mcClient *client.Client, clusterName string, clusterNamespace string) func() error {
 	return func() error {
 		machinePools := &capiexp.MachinePoolList{}
 		machinePoolListOptions := []cr.ListOption{
@@ -272,7 +309,7 @@ func CheckMachinePoolsReadyAndRunning(mcClient *client.Client, clusterName strin
 				"cluster.x-k8s.io/cluster-name": clusterName,
 			},
 		}
-		err := mcClient.List(context.Background(), machinePools, machinePoolListOptions...)
+		err := mcClient.List(ctx, machinePools, machinePoolListOptions...)
 		if err != nil {
 			return err
 		}